feat: #3255 resolving comments
